refactor: rename src/ to lib/ since it's not compiled
