+++ lib/file.js
