+++ lib/generateTasks.js
