+++ lib/getStagedFiles.js
