+++ lib/makeCmdTasks.js
