+++ lib/printErrors.js
