+++ lib/resolveGitDir.js
