+++ lib/resolveTaskFn.js
