+++ lib/validateConfig.js
